AI_image_app/requirements.txt
streamlit
torch
pillow
torchvision
tk